Admin Page (Orders) Functionality
- Admin can now manage orders
- Updated order model
- Updated order and items database functions
- Linked various admin pages to orders
- Added relevant templates
