+++ database/orderitem.go
@@ -47,6 +47,25 @@ func InsertOrderItem(db *sql.DB, myOrderItem models.OrderItem) (err error) {
 	return
 }
 
+// Select/Read a order_item entry from database with a ID input
+func SelectOrderItemByID(db *sql.DB, ID int) (myOrderItem models.OrderItem, err error) {
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	query := "SELECT * FROM order_items WHERE ID=?"
+	err = db.QueryRowContext(ctx, query, ID).Scan(
+		&myOrderItem.ID,
+		&myOrderItem.OrderID,
+		&myOrderItem.FoodID,
+		&myOrderItem.FoodName,
+		&myOrderItem.Quantity,
+		&myOrderItem.SubtotalPrice,
+		&myOrderItem.SubtotalCalories,
+	)
+	return
+}
+
 // Select/Read order_items from database with a orderID input
 func SelectOrderItemsByOrderID(db *sql.DB, orderID int) (myOrderItems []models.OrderItem, err error) {
