+++ templates/admin-restaurants-foods.html
@@ -41,7 +41,7 @@
 			<a
 				class="btn btn-light"
 				style="background-color: #d70f64; color: white"
-				href="/admin/restaurants/{{.Restaurant.ID}}/newfood"
+				href="/admin/restaurants/{{.Restaurant.ID}}/foods/new"
 				role="button"
 			>
 				Create New Food Item
@@ -65,7 +65,7 @@
 			<a
 				class="btn btn-light"
 				style="background-color: #d70f64; color: white"
-				href="#"
+				href="/admin/restaurants/{{.Restaurant.ID}}/orders"
 				role="button"
 			>
 				Manage Orders
@@ -102,7 +102,7 @@
 							<a
 								class="btn btn-light"
 								style="background-color: #d70f64; color: white"
-								href="/admin/restaurants/{{.RestaurantID}}/{{.ID}}"
+								href="/admin/restaurants/{{.RestaurantID}}/foods/{{.ID}}"
 								role="button"
 							>
 								Edit Food
@@ -112,7 +112,7 @@
 							<a
 								class="btn btn-light"
 								style="background-color: #d70f64; color: white"
-								href="/admin/restaurants/{{.RestaurantID}}/{{.ID}}/delete"
+								href="/admin/restaurants/{{.RestaurantID}}/foods/{{.ID}}/delete"
 								role="button"
 								onclick="return confirm('Are you sure you want to delete food item of ID: {{.ID}}?')"
 							>
