+++ templates/admin-restaurants-profile.gohtml
@@ -40,7 +40,7 @@
 			<a
 				class="btn btn-light"
 				style="background-color: #d70f64; color: white"
-				href="/admin/restaurants/{{.ID}}"
+				href="/admin/restaurants/{{.Restaurant.ID}}/foods"
 				role="button"
 			>
 				Manage Food Items
@@ -48,7 +48,7 @@
 			<a
 				class="btn btn-light"
 				style="background-color: #d70f64; color: white"
-				href="#"
+				href="/admin/restaurants/{{.Restaurant.ID}}/orders"
 				role="button"
 			>
 				Manage Orders
