+++ templates/admin-users-profile.html
@@ -39,7 +39,7 @@
 			<a
 				class="btn btn-light"
 				style="background-color: #d70f64; color: white"
-				href="#"
+				href="/admin/users/{{.Target.Username}}/orders"
 				role="button"
 			>
 				Manage Orders
