+++ templates/admin-users.html
@@ -85,7 +85,7 @@
 							<a
 								class="btn btn-light"
 								style="background-color: #d70f64; color: white"
-								href="#"
+								href="/admin/users/{{.Username}}/orders"
 								role="button"
 							>
 								Manage Orders
