+++ templates/admin.html
@@ -19,6 +19,7 @@
 			<h2>Admin Page</h2>
 			<h3><a href="/admin/users">Manage Users</a></h3>
 			<h3><a href="/admin/restaurants">Manage Restaurants</a></h3>
+			<h3><a href="/admin/orders">Manage Orders</a></h3>
 		</div>
 		<!-- Bootstrap Javascript -->
 		{{template "bootstrap-javascript"}}
